requirements.txt
alembic==0.9.7
aniso8601==2.0.0
autoenv==1.0.0
click==6.7
Flask==0.12.2
Flask-JWT==0.3.2
Flask-Login==0.4.1
Flask-Migrate==2.1.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
pbr==3.1.1
pep8==1.7.1
psycopg2==2.7.4
pycrypto==2.6.1
PyJWT==1.4.2
python-dateutil==2.6.1
python-dotenv==0.7.1
python-editor==1.0.3
pytz==2018.3
six==1.11.0
SQLAlchemy==1.2.2
stevedore==1.28.0
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.8.2
Werkzeug==0.14.1
WTForms==2.1
